+++ accessories/humidifier-dehumidifier.js
@@ -86,8 +86,8 @@ class HumidifierDehumidifierAccessory extends FanAccessory {
   async setHumidifierThreshold (hexData, previousValue) {
     const { config, name, log, state, logLevel } = this;
     if (state.HumidifierThreshold === previousValue && config.preventResendHex && !this.previouslyOff) {return;}
-    let desiredState = this.getDesiredState ();
-    let previousState = state.currentState;
+    const desiredState = this.getDesiredState ();
+    const previousState = state.currentState;
     
     if (state.currentState === desiredState) {return;}
     
@@ -99,8 +99,8 @@ class HumidifierDehumidifierAccessory extends FanAccessory {
   async setDehumidifierThreshold (hexData, previousValue) {
     const { config, name, log, state, logLevel } = this;
     if (state.DehumidifierThreshold === previousValue && config.preventResendHex && !this.previouslyOff) {return;}
-    let desiredState = this.getDesiredState ();
-    let previousState = state.currentState;
+    const desiredState = this.getDesiredState ();
+    const previousState = state.currentState;
     
     if (state.currentState === desiredState) {return;}
     
@@ -171,11 +171,11 @@ class HumidifierDehumidifierAccessory extends FanAccessory {
       state.DehumidifierThreshold = 0
     }
     
-    let desiredState = this.getDesiredState ();
+    const desiredState = this.getDesiredState ();
     
     if (state.currentState === desiredState && !this.previouslyOff) {return;}
     
-    let previousState = state.currentState;
+    const previousState = state.currentState;
     if (logLevel <=1) {log(`\x1b[34m[DEBUG]\x1b[0m ${name} updateDeviceState: currently ${state.currentState}, changing to ${desiredState}`);}
 
     state.currentState = desiredState;
@@ -314,7 +314,7 @@ class HumidifierDehumidifierAccessory extends FanAccessory {
       } else {
         lines.forEach((line) => {
           if(-1 < line.indexOf(':')){
-            let value = line.split(':');
+            const value = line.split(':');
             if(value[0] == 'temperature') {temperature = parseFloat(value[1]);}
             if(value[0] == 'humidity') {humidity = parseFloat(value[1]);}
             if(value[0] == 'battery' && batteryAlerts) {state.batteryLevel = parseFloat(value[1]);}