+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.4.20 - 2025-08-28]
+### Fixed
+ - Fixed ESLint support and applied some code fixes
+### Changed
+ - Stopped device temperature checks when they're not applicable (MQTT, File, W1)
+ - Updated dependancies
+
 ## [4.4.19 - 2024-09-30]
 ### Fixed
  - refreshCharacteristic and value() were not updating the UI. Rebuilt to use updateValue method. (#733 and #725)